+++ src/Components/Patient/PatientOverviewComponent.tsx
@@ -1,10 +1,13 @@
-import { Patient } from "fhir/r4";
+import { MedicationStatement, Patient } from "fhir/r4";
 import PatientGeneralWidgetComponent from "./PatientGeneralWidgetComponent";
 import ObservationService from "../../Services/ObservationService";
 import InfoListComponent, { InfoListData } from "../InfoListComponent";
 import { useEffect, useState } from "react";
 import ConditionService from "../../Services/ConditionService";
-import MedicationService from "../../Services/MedicationService";
+import ConditionUtils from "../../Services/Utils/ConditionUtils";
+import ObservationUtils from "../../Services/Utils/ObservationUtils";
+import FhirResourceService from "../../Services/FhirService";
+import MedicationUtils from "../../Services/Utils/MedicationUtils";
 export default function PatientOverviewComponent({
   patient,
 }: {
@@ -12,22 +15,18 @@ export default function PatientOverviewComponent({
 }) {
   const observationService = new ObservationService();
   const conditionService = new ConditionService();
-  const medicationService = new MedicationService();
+  const medicationService = new FhirResourceService("MedicationStatement");
 
   const [observationData, setObservationData] = useState<InfoListData[]>([]);
   const [conditionData, setConditionData] = useState<InfoListData[]>([]);
   const [medication, setMedicationData] = useState<InfoListData[]>([]);
-  
-
 
   const fetchObservationData = async () => {
     const result = await observationService.getResources({
       subject: patient.id!,
     });
     if (result.success) {
-      setObservationData(
-        observationService.extractObservationInfo(result.data)
-      );
+      setObservationData(ObservationUtils.extractObservationInfo(result.data));
       console.log(result.data);
     }
   };
@@ -37,7 +36,7 @@ export default function PatientOverviewComponent({
       subject: patient.id!,
     });
     if (result.success) {
-      setConditionData(conditionService.extractConditionName(result.data));
+      setConditionData(ConditionUtils.extractConditionName(result.data));
       console.log(result.data);
     }
   };
@@ -47,7 +46,11 @@ export default function PatientOverviewComponent({
       subject: patient.id!,
     });
     if (result.success) {
-      setMedicationData(medicationService.extractMedicationInfo(result.data));
+      setMedicationData(
+        MedicationUtils.extractMedicationInfo(
+          result.data as MedicationStatement[]
+        )
+      );
       //setConditionData(conditionService.extractConditionName(result.data));
       console.log("medication");
       console.log(result.data);
@@ -62,7 +65,6 @@ export default function PatientOverviewComponent({
     fetchConditionData();
     fetchMedicationData();
   }, [patient.id]);
-
 
   return (
     <div style={{ padding: "50px" }}>
@@ -79,7 +81,6 @@ export default function PatientOverviewComponent({
             icon={"/hearth.svg"}
             resourceType="Observation"
             patientId={patient.id!}
-            
           ></InfoListComponent>
         </div>
         <div style={{ flex: 1 }}>
