+++ src/Components/Patient/PatientQuestionnaireComponent.tsx
@@ -1,15 +1,16 @@
 import { useEffect, useState } from "react";
 import { Questionnaire, QuestionnaireResponse } from "fhir/r4";
-import QuestionnaireResponseService from "../../Services/QuestionnaireResponseService";
 import QuestionnaireComponent from "../Questionnaire/QuestionnaireComponent";
 import QuestionnaireListComponent from "../Questionnaire/QuestionnaireListDialogComponent";
 import { Accordion, AccordionDetails, AccordionSummary } from "@mui/material";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import QuestionnaireService from "../../Services/QuestionnaireService";
 import { isAdminOrPractitioner } from "../../RolUser";
+import FhirResourceService from "../../Services/FhirService";
 
-const questionnaireResponseService = new QuestionnaireResponseService();
-const questionnaireService = new QuestionnaireService();
+const questionnaireResponseService = new FhirResourceService(
+  "QuestionnaireResponse"
+);
+const questionnaireService = new FhirResourceService("Questionnaire");
 
 export default function PatientQuestionnaireComponent({
   patientID,
@@ -42,15 +43,16 @@ export default function PatientQuestionnaireComponent({
       if (!responseBundle.success) throw Error(responseBundle.error);
 
       console.log(responseBundle.data);
-      setQuestionnaireResponses(responseBundle.data);
+      setQuestionnaireResponses(responseBundle.data as QuestionnaireResponse[]);
       const updatedQuestionnaires: Record<string, Questionnaire> = {};
 
-      for (const quetionnaireResponse of responseBundle.data) {
+      for (const quetionnaireResponse of responseBundle.data as QuestionnaireResponse[]) {
         const quesR_id = quetionnaireResponse.questionnaire;
         if (!quesR_id) continue;
 
         const res = await questionnaireService.getById(quesR_id);
-        if (res.success) updatedQuestionnaires[quesR_id] = res.data;
+        if (res.success)
+          updatedQuestionnaires[quesR_id] = res.data as Questionnaire;
       }
       setQuestionnaires(updatedQuestionnaires);
     } catch {
