+++ src/Pages/ListPages/EncounterPatientPage.tsx
@@ -8,14 +8,17 @@ import { useParams } from "react-router-dom";
 import ConditionService from "../../Services/ConditionService";
 import PatientQuestionnaireComponent from "../../Components/Patient/PatientQuestionnaireComponent";
 
-export default function EncounterPatientPage() {
-  const observationService = new ObservationService();
-  const conditionService = new ConditionService();
+import ObservationUtil from "../../Services/Utils/ObservationUtils";
+import ConditionUtils from "../../Services/Utils/ConditionUtils";
 
+const observationService = new ObservationService();
+const conditionService = new ConditionService();
+
+export default function EncounterPatientPage() {
   const { patientID } = useParams();
   const { encounterID } = useParams();
 
-  const [obvservationData, setObvservationData] = useState<InfoListData[]>([]);
+  const [observationData, setObservationData] = useState<InfoListData[]>([]);
 
   const [conditionData, setConditionData] = useState<InfoListData[]>([]);
 
@@ -28,9 +31,7 @@ export default function EncounterPatientPage() {
     console.log("patientID ", patientID);
     console.log("encounterID ", encounterID);
     if (result.success) {
-      setObvservationData(
-        observationService.extractObservationInfo(result.data)
-      );
+      setObservationData(ObservationUtil.extractObservationInfo(result.data));
       console.log("observation:", result.data);
     }
 
@@ -40,25 +41,25 @@ export default function EncounterPatientPage() {
     });
     if (resultCon.success) {
       console.log("conditions: ", resultCon.data);
-      setConditionData(conditionService.extractConditionName(resultCon.data));
+      setConditionData(ConditionUtils.extractConditionName(resultCon.data));
       //console.log(conditionService.extractConditionName(resultCon.data));
     }
   };
 
   useEffect(() => {
-    setObvservationData([]);
+    setObservationData([]);
     fetchData();
   }, [patientID]);
 
   return (
     <div style={{ padding: "50px" }}>
       <div style={{ paddingBottom: "30px" }}>
-        <h1>info encountro</h1>
+        <h1>info encuentro</h1>
       </div>
       <div style={{ display: "flex", flexWrap: "wrap", gap: "30px" }}>
         <div style={{ flex: 1 }}>
           <InfoListComponent
-            data={obvservationData}
+            data={observationData}
             title={"Observaciones"}
             icon={"/hearth.svg"}
             resourceType="Observation"
