+++ src/Services/FhirService.ts
@@ -8,6 +8,10 @@ import {
   CodeableConcept,
 } from "fhir/r4";
 
+/**
+ * Represents a service for interacting with FHIR resources.
+ * @template T - The type of FHIR resource.
+ */
 export default class FhirResourceService<T extends FhirResource> {
   public resourceTypeName: string;
 
@@ -19,6 +23,10 @@ export default class FhirResourceService<T extends FhirResource> {
   public hasNextPage: boolean = false;
   public hasPrevPage: boolean = false;
 
+  /**
+   * Creates an instance of ResourceService.
+   * @param {FhirType} type - The FHIR resource type.
+   */
   public constructor(type: FhirType) {
     this.apiUrl =
       import.meta.env.VITE_API_URL || "https://hapi.fhir.org/baseR4";
@@ -27,6 +35,13 @@ export default class FhirResourceService<T extends FhirResource> {
     this.resourceTypeName = type;
   }
 
+  /**
+   * Handles the result of a FHIR operation.
+   * @private
+   * @template T - The type of the result.
+   * @param {Promise<T>} result - The result of the FHIR operation.
+   * @returns {Promise<Result<T>>} - The handled result.
+   */
   private handleResult<T>(result: Promise<T>): Promise<Result<T>> {
     return result
       .then((data: T) => {
@@ -42,6 +57,11 @@ export default class FhirResourceService<T extends FhirResource> {
       });
   }
 
+  /**
+   * Sets the flags indicating whether there is a next or previous page of resources.
+   * @private
+   * @param {Bundle} response - The FHIR bundle response.
+   */
   private setNewPages(response: Bundle) {
     this.hasNextPage = !!response.link?.find((link) => link.relation === "next")
       ?.url;
@@ -50,6 +70,11 @@ export default class FhirResourceService<T extends FhirResource> {
     )?.url;
   }
 
+  /**
+   * Retrieves a FHIR resource by its ID.
+   * @param {string} id - The ID of the resource.
+   * @returns {Promise<Result<T>>} - The result of the operation.
+   */
   public async getById(id: string): Promise<Result<T>> {
     return this.handleResult<T>(
       this.fhirClient.read({
@@ -59,6 +84,11 @@ export default class FhirResourceService<T extends FhirResource> {
     );
   }
 
+  /**
+   * Retrieves the history of a FHIR resource by its ID.
+   * @param {string} id - The ID of the resource.
+   * @returns {Promise<Result<T[]>>} - The result of the operation.
+   */
   public async getHistoryById(id: string): Promise<Result<T[]>> {
     const result = await this.handleResult<Bundle>(
       this.fhirClient.history({
@@ -76,6 +106,11 @@ export default class FhirResourceService<T extends FhirResource> {
     }
   }
 
+  /**
+   * Creates a new FHIR resource.
+   * @param {T} newResource - The new resource to create.
+   * @returns {Promise<Result<T>>} - The result of the operation.
+   */
   public async postResource(newResource: T): Promise<Result<T>> {
     return this.handleResult<T>(
       this.fhirClient.create({
@@ -85,6 +120,12 @@ export default class FhirResourceService<T extends FhirResource> {
     );
   }
 
+  /**
+   * Determines the action to perform on a resource in a bundle.
+   * @private
+   * @param {any} objeto - The resource object.
+   * @returns {"POST" | "PUT" | "DELETE" | "GET" | "PATCH"} - The action to perform.
+   */
   private bundleAction(
     objeto: any
   ): "POST" | "PUT" | "DELETE" | "GET" | "PATCH" {
@@ -108,6 +149,14 @@ export default class FhirResourceService<T extends FhirResource> {
     else if (objeto.id) return "PUT";
     else return "POST";
   }
+
+  /**
+   * Generates the URL for a resource based on the method.
+   * @private
+   * @param {Resource} resource - The FHIR resource.
+   * @param {"POST" | "PUT" | "DELETE" | "GET" | "PATCH"} method - The HTTP method.
+   * @returns {string} - The resource URL.
+   */
   private resourceUrl(
     resource: FhirResource,
     method: "POST" | "PUT" | "DELETE" | "GET" | "PATCH"
@@ -118,16 +167,9 @@ export default class FhirResourceService<T extends FhirResource> {
   }
 
   /**
-   *
-   * @param newResources Arreglo del recurso,
-   *  este se encarga de enviarlo como un bundle.
-   *
-   * Condiciones:
-   *
-   * sólo @id y @resourceType elimina el recurso,
-   * si tiene @id actualiza el recurso,
-   * en caso de que el @id no existe, lo añade.
-   * @returns
+   * Sends an array of resources as a bundle.
+   * @param {T[]} newResources - The array of resources to send.
+   * @returns {Promise<Result<T>>} - The result of the operation.
    */
   public async sendArray(newResources: T[]): Promise<Result<T>> {
     const result = await this.handleResult(
@@ -154,6 +196,11 @@ export default class FhirResourceService<T extends FhirResource> {
     }
   }
 
+  /**
+   * Updates an existing FHIR resource.
+   * @param {T} newResource - The updated resource.
+   * @returns {Promise<Result<T>>} - The result of the operation.
+   */
   public async updateResource(newResource: T): Promise<Result<T>> {
     return this.handleResult<T>(
       this.fhirClient.update({
@@ -164,13 +211,22 @@ export default class FhirResourceService<T extends FhirResource> {
     );
   }
 
+  /**
+   * Sends a FHIR resource. If the resource has an ID, it updates the resource; otherwise, it creates a new resource.
+   * @param {T} newResource - The resource to send.
+   * @returns {Promise<Result<T>>} - The result of the operation.
+   */
   public async sendResource(newResource: T): Promise<Result<T>> {
     return newResource.id
       ? this.updateResource(newResource)
       : this.postResource(newResource);
   }
 
-  //TODO: Verificar el count para posibles recursos donde pueden ser muchos y no he aplicado el NextPage o PrevPage
+  /**
+   * Retrieves a list of FHIR resources.
+   * @param {SearchParams} [params] - The search parameters.
+   * @returns {Promise<Result<T[]>>} - The result of the operation.
+   */
   public async getResources(params?: SearchParams): Promise<Result<T[]>> {
     const result = await this.handleResult<Bundle>(
       this.fhirClient.search({
@@ -195,6 +251,11 @@ export default class FhirResourceService<T extends FhirResource> {
     }
   }
 
+  /**
+   * Retrieves the next or previous page of resources.
+   * @param {"next" | "prev"} direction - The direction of the page to retrieve.
+   * @returns {Promise<Result<T[]>>} - The result of the operation.
+   */
   public async getNewResources(
     direction: "next" | "prev"
   ): Promise<Result<T[]>> {
@@ -218,10 +279,22 @@ export default class FhirResourceService<T extends FhirResource> {
     return { success: true, data: resources };
   }
 
+  /**
+   * Parses the operation outcome to extract the error message.
+   * @private
+   * @param {OperationOutcome} operation - The operation outcome.
+   * @returns {string} - The error message.
+   */
   private parseOperationOutcome(operation: OperationOutcome): string {
     return operation.issue[0]?.diagnostics || "Unknown error";
   }
 
+  /**
+   * Retrieves the coding section of a CodeableConcept by system.
+   * @param {CodeableConcept} code - The CodeableConcept.
+   * @param {string} system - The coding system.
+   * @returns {CodeableConcept | null} - The coding section or null if not found.
+   */
   public getCodingBySystem(code: CodeableConcept, system: string) {
     const coding = code.coding;
     if (!coding || coding.length === 0) return null;
