--- src/Services/QuestionnaireResponseService.ts
@@ -1,10 +1,0 @@
-import { QuestionnaireResponse } from "fhir/r4";
-import FhirResourceService from "./FhirService";
-
-export default class QuestionnaireService extends FhirResourceService<QuestionnaireResponse> {
-    
-    constructor() {
-      super("QuestionnaireResponse");
-    }
-
-}
