--- src/Services/QuestionnaireService.ts
@@ -1,8 +1,0 @@
-import { Questionnaire} from "fhir/r4";
-import FhirResourceService from "./FhirService";
-
-export default class QuestionnaireService extends FhirResourceService<Questionnaire> {
-  constructor() {
-    super("Questionnaire");
-  }
-}
