+++ src/Services/Utils/ConditionUtils.ts
@@ -1,0 +1,27 @@
+import { Condition } from "fhir/r4";
+import { InfoListData } from "../../Components/InfoListComponent";
+
+export default class ConditionUtils {
+    public static extractConditionName(conditions: Condition[]): InfoListData[] {
+        return conditions.map((condition) => {
+            const id = condition.id;
+            const name =
+                condition.code?.coding?.[0]?.display ||
+                condition.code?.text ||
+                (condition.code?.coding?.[0]?.system &&
+                    condition.code?.coding?.[0]?.code
+                    ? `${condition.code.coding[0].system} - ${condition.code.coding[0].code}`
+                    : "Unknown Name");
+
+            const value =
+                condition.clinicalStatus?.coding?.[0]?.display ||
+                condition.clinicalStatus?.text ||
+                (condition.clinicalStatus?.coding?.[0]?.system &&
+                    condition.clinicalStatus?.coding?.[0]?.code
+                    ? `${condition.clinicalStatus.coding[0].code}`
+                    : "Unknown Name");
+
+            return { id, name, value };
+        });
+    }
+}
